requirements.txt
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
rapidfuzz==3.10.1
PyYAML==6.0.2
pytest==8.3.3
